Move schema to the root to make it more visible 👀
